feat(grammar): Makes `,` an available symbol for binary messages
This promotes the previously special `,` symbol to an available symbol
for binary messages. Previously this symbol was used to separate
expressions in some context. Those contexts can use the `;` symbol
instead, which is already used as an expression separator in other
places.

The `,` message makes a lot of sense in contexts like parser
combinators:

   def number = this digit, this digit many

Furthermore, this change defines `,` and `.` as separator messages,
which means that they may not be part of an unary or keyword identifier
name. This is necessary so the above example doesn't get parsed as:

  number: send(...send(send(this, 'digit,'), 'this'...)