feat(semantics): makes all modules parametric
Previously the generated code relied on a global `Siren` object. While
this wasn't much of an issue executing Siren files directly, but
compiling modules and using them later was a major pain of implicit
dependencies.

Now all files are compiled as parametric modules, with the first
argument being the runtime. When loading Siren modules from the JS side,
one's expected to pass the runtime along.