refactor(runtime): New stdlib organisation w/ better JS integration & BigInts
This refactoring in the runtime and standard library brings a better JS
integration, by introducing the concept of JS Alien objects, which proxy
communication between Siren and JS land. These allow Siren code that
interacts with JS to look very similar to JS code itself:

    let Number = JS-Alien global at: "Number";
    let a = Number at: "MIN_SAFE_INTEGER";
    let b = Number at: "MAX_SAFE_INTEGER";
    JS-Alien global at: "console" >> invoke: "log" with: (a < b)

I'm still considering whether to use does-not-understand:with-arguments:
to bridge the gap even more, such that you could instead write:

    let Number = JS-Alien global Number;
    let a = Number MIN_SAFE_INTEGER;
    let b = Number MAX_SAFE_INTEGER;
    JS-Alien global console log: (a < b)

Since the alien provides a few methods out of the box, this would of
course be usable only for a subset of properties in the proxied JS
object, but it makes working with JS values much closer to how one would
work with Siren values (Newspeak does this, by the way).

- - -

This also rethinks the core types.
These types are now factored out into "internal libraries". Each
internal library is a file, preceded with its "loading priority".
So the Core library is named `00-Core.siren`, because it has to
be the first thing to be loaded.

The following are the new core types in Siren:

- Result, used for representing the return of computations that can fail.
- Block, used for representing executable code.
- Boolean, used for representing boolean logic.
- String, used for representing text.
- Tuple, used for representing sequences of values.
- Float, used for representing floating-point numbers.
- Integer, used for representing arbitrary-precision integers.
- Reference, used for representing mutable slots.

With the exception of `Reference`, all of the core types
are *immutable*.

Block and Reference use reference equality, whereas all other core types
use structural equality.

Note that this refactoring also introduces big
integers. Arbitrary-precision integers are provided with
the object Integer, and JS native floating point values
keep using the Float object.

This fixes #29.

Additionally, the type of the number is selected from
its notation:

    1000   => Integer
    1000.0 => Float