feat(grammar): Make `this` an explicit argument
BREAKING CHANGE:

Now Siren has explicit receiver arguments. This avoids problems with
object nesting, at the cost of some verbosity. With the new `_` for
ignoring arguments, the verbosity shouldn't be too much of a problem.

Where one used to write:

    let pair = {
      def + that {
        let receiver = this;
        this {
          def x = receiver x + that x
          def y = receiver y + that y
        }
      }
    }

They'd now write:

     let pair = {
       def this + that
         this {
           def _ x = this x + that x
           def _ y = this y + that y
         }
     }

`this` isn't a special identifier anymore. And `_` can be used as a
special symbol in place of arguments for the ones you don't have any
interest in.