feat(Concurrency): Task-based concurrency
This introduces a concurrency primitive for actions based on a Task
monad.

Tasks represent an action (an asynchronous computation). Users
can run Tasks just as they can invoke functions. Every time one runs a
Task they receive back a `Task-Execution` object, which represents the
execution of the task. `Task-Execution` objects know how to cancel the
execution of the task that they represent, and can also provide a
`Future` for the result of that task.

Furthermore, Tasks can be composed. The `||` operator represents a
non-deterministic choice between two tasks, whereas the `,` operator
represents the deterministic concatenation of two tasks.