feat(Runtime): Uses @name meta for describe by default
This change makes objects use their `name` meta-data by default for
describing themselves (in debug). This gives you somewhat relevant
information even if you forget to define a `describe` message in it, as
long as the object has a `name` meta-data.

Given that Siren propagates naming meta-data automatically when it can
infer it, this works nicely:

    let Hello = { def self hello: name = "Hello ", name, "!". }
    # => <Hello>

Anonymous objects are still described uselessly as "Anonymous Object":

    { def self hello: name = "Hello ", name, "!". }
    # => <Anonymous Object>

Once #52 is implemented, this information can be substantially
improved.