fix(Codegen): Updates use/using to use perspectives
BREAKING CHANGE:

Use/Using weren't updated to use the new Perspectives object when the
runtime was rewritten. This commit moves use/using forms to this new
runtime, and consequently removes the multi-traits special form, which
is now reified in the Perspective object itself.

So, where one used to write:

    use trait-a. trait-b in { ... }

They'd now write:

    use trait-a, trait-b in { ... }

Where `,` is the concatenation message understood by the Perspective
object.